Added module docstring.
The __version__ and __author__ variables should not live inside a
docstring but rather at module level, they are interpreted by pydoc
(e.g., python's help() function).

The `#!` line as well as the encoding line should be the first lines of
a file in order to be recognized.

If the first statement is a string literal, it will be used as a
module-level docstring.